:recycle:refactor: seperate user menu from header 헤더의 유저 메뉴 분리
