+++ src/components/Header/Header.tsx
@@ -1,12 +1,10 @@
-import React, { useState, KeyboardEvent } from 'react';
+import { useState, KeyboardEvent } from 'react';
 import { useNavigate } from 'react-router-dom';
 import SearchIcon from '@mui/icons-material/Search';
-import PersonIcon from '@mui/icons-material/Person';
 import { useTheme } from '@mui/material/styles';
-import { Box, Button, Menu, MenuItem, InputBase } from '@mui/material';
-import { useSelector, useDispatch } from 'react-redux';
+import { Box, Button, InputBase } from '@mui/material';
+import { useSelector } from 'react-redux';
 import { RootState } from '@store/index';
-import { logoutSuccess } from '@features/user/userSlice';
 import DarkModeButton from '@components/DarkModeButton/DarkModeButton';
 
 import {
@@ -15,17 +13,14 @@ import {
   StyledSearchContainer,
   StyledIconWrapper,
 } from './Header.styles';
-import { supabase } from '@utils/supabaseClient';
+import UserMenu from './userMenu';
 
 const Header = (): JSX.Element => {
   const [showSearchBar, setShowSearchBar] = useState(false);
   const [searchQuery, setSearchQuery] = useState('');
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
   const user = useSelector((state: RootState) => state.user);
 
   const navigate = useNavigate();
-  const dispatch = useDispatch();
-  const open = Boolean(anchorEl);
   const theme = useTheme();
 
   const toggleSearchBar = () => {
@@ -36,53 +31,20 @@ const Header = (): JSX.Element => {
   };
 
   const handleSearch = () => {
-    const trimmedQuery = searchQuery.trim(); // 공백 제거
-
+    const trimmedQuery = searchQuery.trim();
     if (trimmedQuery) {
       navigate(`/search?query=${encodeURIComponent(trimmedQuery)}`);
-      setSearchQuery(''); // 검색어 초기화
-      setShowSearchBar(false); // 검색창 닫기
+      setSearchQuery('');
+      setShowSearchBar(false);
     } else {
       navigate('/search');
-      setShowSearchBar(false); // 검색창 닫기
+      setShowSearchBar(false);
     }
   };
 
   const handleKeyPress = (event: KeyboardEvent<HTMLInputElement>) => {
     if (event.key === 'Enter') {
       handleSearch();
-    }
-  };
-
-  const handleProfileClick = (event: React.MouseEvent<HTMLDivElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
-  const handleNavigation = (path: string) => {
-    navigate(path);
-    handleClose();
-  };
-
-  const handleLogout = async () => {
-    try {
-      // Supabase 로그아웃
-      await supabase.auth.signOut();
-
-      // Redux 상태 업데이트
-      dispatch(logoutSuccess());
-
-      // 로컬 스토리지에서 토큰 제거
-      localStorage.removeItem('token');
-
-      handleClose(); // 메뉴 닫기
-      navigate('/'); // 홈으로 이동
-    } catch (error) {
-      console.error('로그아웃 중 오류 발생:', error);
-      // 오류 처리 (예: 사용자에게 알림)
     }
   };
 
@@ -111,32 +73,10 @@ const Header = (): JSX.Element => {
           </StyledIconWrapper>
         </StyledSearchContainer>
         {user.isLoggedIn ? (
-          <div>
-            <StyledIconWrapper onClick={handleProfileClick}>
-              <PersonIcon />
-            </StyledIconWrapper>
-            <Menu
-              id="basic-menu"
-              anchorEl={anchorEl}
-              open={open}
-              onClose={handleClose}
-              aria-hidden="false"
-              MenuListProps={{
-                'aria-labelledby': 'basic-button',
-              }}
-            >
-              <MenuItem onClick={() => handleNavigation('/my-page')}>
-                마이페이지
-              </MenuItem>
-              <MenuItem onClick={() => handleNavigation('/edit-account')}>
-                개인정보수정
-              </MenuItem>
-              <MenuItem onClick={handleLogout}>로그아웃</MenuItem>
-            </Menu>
-          </div>
+          <UserMenu />
         ) : (
           <Button
-            onClick={() => handleNavigation('/login')}
+            onClick={() => navigate('/login')}
             variant="outlined"
             color="primary"
           >
